+++ packages/nx-rust/src/generators/library/generator.ts
@@ -1,6 +1,10 @@
-import * as nrwl from '@nrwl/devkit';
-import { Tree } from '@nrwl/devkit';
-import * as path from 'path';
+import {
+  addProjectConfiguration,
+  formatFiles,
+  generateFiles,
+  Tree,
+} from '@nrwl/devkit';
+import { join } from 'path';
 
 import {
   GeneratorOptions,
@@ -11,13 +15,11 @@ import { CARGO_TOML, PLUGIN_NAME } from '../../common/constants';
 import cargoInit from '../init/generator';
 import CLIOptions from './schema';
 
-// prettier-ignore
 type Options = CLIOptions & GeneratorOptions;
 
 export default async function (host: Tree, opts: CLIOptions) {
   const options = normalizeGeneratorOptions('library', host, opts);
-
-  nrwl.addProjectConfiguration(host, options.projectName, {
+  addProjectConfiguration(host, options.projectName, {
     root: options.projectRoot,
     projectType: 'library',
     sourceRoot: `${options.projectRoot}/src`,
@@ -28,36 +30,27 @@ export default async function (host: Tree, opts: CLIOptions) {
       },
       lint: {
         executor: `${PLUGIN_NAME}:clippy`,
-        options: {
-          fix: false,
-          failOnWarnings: true,
-          noDeps: true,
-        },
+        options: { fix: false, failOnWarnings: true, noDeps: true },
       },
     },
     tags: options.parsedTags,
   });
-
   await addFiles(host, options);
   updateWorkspaceMembers(host, options);
-  await nrwl.formatFiles(host);
+  await formatFiles(host);
 }
 
 async function addFiles(host: Tree, opts: Options) {
-  if (!host.exists(CARGO_TOML)) {
-    await cargoInit(host, {});
-  }
-
+  if (!host.exists(CARGO_TOML)) await cargoInit(host, {});
   const substitutions = {
     projectName: opts.projectName,
     moduleName: opts.moduleName,
     edition: opts.edition,
     template: '',
   };
-
-  nrwl.generateFiles(
+  generateFiles(
     host,
-    path.join(__dirname, 'files'),
+    join(__dirname, 'files'),
     opts.projectRoot,
     substitutions
   );